bot/pyproject.toml
[project]
name = "linksaver-bot"
version = "1.0.0"
description = "LinkSaver Telegram Bot with LLM integration"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.14.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
]
